Add support for building `protobuf-wire` via `nix`
The motivation for this is to move the `nix` build logic for each Haskell
package out of the end-to-end setup and into each respective repository so that
each repository can maintain their respective Nix expression.  This greatly
simplifies maintenance for each respective repository and the end-to-end setup
because:

* repositories can locally verify they build correctly via `nix` without running
  their branch through the end-to-end setup
* the end-to-end setup only needs to update the git revision when upgrading a
  repository (instead of also updating the `nix` expression, too)

This also enables testing this repository via Hydra although I haven't
implemented that yet.