Fix `protobuf-wire` to build with `nix` (#59)
* Fix `protobuf-wire` to build with `nix`

This requires a somewhat complicated setup due to the fact that the test suite
requires the `protobuf-wire` Haskell package to be available as a Haskell
dependency, but the derivations that `cabal2nix` generates don't support this.

I worked around this by:

* creating a `protobuf-wire-no-tests` derivation based off of `default.nix`
  built without tests
* using that as an input to the `protobuf-wire` derivation based off of
  `default-tests.nix` built with tests

The `default.nix` is still the recommended import for other Haskell packages
that depend on `protobuf-wire`.  It's just that they won't run the tests upon
importing our library.  Instead, they just rely on the fact that our CI runs the
tests on each pull request.

There were two other small changes that were necessary to get `nix` to run the
tests successfully:

* Update the tests to append to the `PYTHONPATH` instead of replacing it
* Adding a `nix`-specific patch to use `ghc` instead of `stack ghc` in the tests

I also updated the `proto3-wire` revision to match the one in the `stack.yaml`
file