Simplify Nix logic
This simplifies the logic in `release.nix` by taking advantage of the
`pkgs.haskell.lib.*` suite of utilities.  The main changes are:

* Remove `default-tests.nix` and use `default.nix` exclusively
* Refactor `proto3-suite-no-tests` and `proto3-suite` to use `default.nix`
* Pin to specific `nixpkgs` revision
* Use the default Haskell compiler (i.e. `ghc-8.0.2`) instead of `ghc-7.10.3`

The main benefit of this change is that updates to `default.nix` are
much simpler.  You can now just run `cabal2nix . > default.nix` and
you no longer need to manually post-process the output.