Transform QuickCheck code into `Arbitrary` instances
The `Proto3.Suite.DotProto.AST` module has several `arbitraryXXX`
functions at the bottom of the module which could be made into
`Arbitrary` instances for their respective types.  This change
refactors the module to create `Arbitrary` instances for as many
types as possible.

This refactor is behavior-preserving, meaning that some of these
`Arbitrary` instances are not as arbitrary as they could be.  For
example, some of the instances return fields that are always an
empty string when they could be arbitrary strings.  This is because
this was the original behavior of the corresponding `arbitraryXXX`
function.

I also try to avoid using operators as much as possible and use `do`
notation wherever possible.