Support `--includeDir` switches on `compile-proto-file` and `canonicalize-proto-file`
This commit adds the notion of an include search path to the
`Proto3.Suite.DotProto.Generate.readDotProtoWithContext` function. The search
path is simply a list of paths which are prepended in sequence to any .proto
file included when compiling a toplevel .proto file.

Also adds dependencies on the `foldl` and `neat-interpolation` packages; the
former is used for convenience, while the latter is used for its pleasant
construction of detailed error messages.