Partial fix for sfixed32/sfixed64 bug in `{To,From}JSONPB` instances
The actual datatype generated for e.g. Scalar32 does not wrap fields of protobuf
type sfixed32 with the Signed wrapper in the type definition, rather handling
its un/wrapping in the de/encodeMessage. This provides a convenient interface
for the user of the datatype, but means that we need to do the un/wrapping
ourselves in the JSONPB instances.

This gets rid of the old and incorrect PLACEHOLDER INSTANCES for (Fixed Int32)
and (Fixed Int64).