Fix pattern binder order bug in JSONPB instance generation for oneof
We were incorrectly ordering pattern binders: pattern binders for normal fields
came first, followed by pattern binders for oneof fields. The correct behavior
is to order the pattern binders in syntactic order of their occurrence in the
protobuf type (which is the same as the record field order in the corresponding
generated Haskell type).

Also includes some misc improvements and some light refactoring and renames.