Provide a `toJSONPB` analog to `Data.Aeson.toJSON` (#37)
The preliminary JSONPB implementation chose to only provide the
`toEncoding`-like interface, because we didn't think that there was a need to
render protobuf values to the `Data.Aeson.Value` intermediate representation.

However, it's become apparent that this was an oversight, as we would like to
generate `aeson` instances in terms of the JSONPB typeclasses.

This commit adds the `toJSONPB` function to the `ToJSONPB` class and makes
corresponding changes to code generation to provide implementations of the new
function.

NB: Due to this change, there's now some unfortunate code duplication, since the
generated implementations for `toJSONPB` and `toEncodingPB` are nearly
identical. However, we choose not to address this at this time; we'd like to
switch to a `Generic`-based implementation eventually, and so are content with
the minimum viable functionality for the time being.