Allow packed repeated fields of enums
Looks like we've never explicitly tested support for packed repeated
enum fields. Doh.

Prior to this commit, the incorrect behavior was as follows: assume
some protobuf enum `E`, and a field definition such as:

```
repeated E enums = 1;
```

(without an explicit `[packed=true]` option in the AST).

In this scenario, the code generator consulted the `isPackableType`
predicate, which stated that all `Named` primitive types are
unpackable. But only message types are unpackable; enums (also
represented by a `Named` primtype) are encoded as varints and may be
packed.

This commit changes the logic used to determine whether or not a type
is packable. The predicate has been renamed to simply `isPackable`,
moved into the code generation module, and now makes use of the code
generator's `TypeContext` information to distinguish messages from
enums. This change to the predicate in turn fixes the related
generated code.

Finally, this commit tweaks the defs of `wrapE` and `unwrapE`
slightly, to taste.