Directly generate `ToSchema` instance from `.proto` AST (#63)
The motivation for this change is to fix the mismatch between the Swagger API
declared in `ToSchema` instances and the actual API served by the `FromJSONPB`
and `ToJSONPB` instances.

The difference between the two was in casing conventions: the JSONPB API was
preserving the original case of the fields from the `.proto` API, whereas the
Swagger API was camel-casing the declared fields.

The reason they differ was that they were going through two separate code
pathways.  In the case of JSONPB the `FromJSONPB` and `ToJSONPB` instances were
being generated directly from the `.proto` AST without any intermediate steps:

    `.proto` AST →(code generation)→ `FromJSONPB`

    `.proto` AST →(code generation)→ `ToJSONPB`

... and that code generation step preserved the original case.

However, the `ToSchema` instance was generated indirectly via GHC generics:

    `.proto` AST →(code generation)→ Haskell data type →(Generics)→ `ToSchema`

... and the code generation step for the Haskell data type was *not* preserving
case.

After this change the code generation directly generates the `ToSchema`
instance:

    `.proto` AST →(code generation)→ `ToSchema`

Some alternative approaches that I considered and rejected:

* Change the Haskell data type generation to preserve the original case instead
  of camel-casing Haskell field names

  I rejected this approach because it is highly disruptive to downstream code
  and also generates ugly field names

* Customize the `ToSchema` instances derived via `Generic`s to convert camel
  case to snake case

  I rejected this approach because it is not correct if the original `.proto`
  field names were already camel-cased.  This would then lead to the opposite
  problem: declaring a snake-cased field which was actually camel-cased

I tested this downstream on our large internal `.proto` files.  The only
difference in the generated Swagger is that the casing for declared fields is
fixed and the generated Haskell code still compiles without any changes to the
client packages.

I also added some simple golden tests for simple visual inspection of what the
generated Swagger looks like.