Make Dhall optional with a cabal configure flag (by default off)
This change also refactors the `release.nix` to automatically produce
the necessary haskell source files from test protos to run the test
suite.

The motivation for this change came from making Dhall optional,
running tests for both the dhall and non-dhall derivations would not
work if we continued to check-in the code generated artifacts.