Add optional Dhall.Interpret and Dhall.Inject instances for generated protobuf Haskell types (#64)
* Generate instances for Dhall.Interpret (using Generic)

* Adding more Dhall.Interpret instances

* Add unsafe Dhall.Interpret instances for Strict and Lazy ByteStrings

If the interpret instances only decode Dhall rendered from the same
types then no unexpected errors should occur but there exists a
possiblity that transformation of the rendered Dhall data might happen
which introduces the possibility of unexpected errors thrown to the
user.

* Add Dhall.Inject instances and better documentation

* Generate instance decls for OneOfs too, add a round trip test

* Make Dhall optional with a cabal configure flag (by default off)

This change also refactors the `release.nix` to automatically produce
the necessary haskell source files from test protos to run the test
suite.

The motivation for this change came from making Dhall optional,
running tests for both the dhall and non-dhall derivations would not
work if we continued to check-in the code generated artifacts.

* Remove unused test-files/simple.proto

* Add `Manual: True` to the dhall cabal flag definition

... as suggested by @Gabriel439.

* Generate a separate Haskell package derivation with dhall enabled

Instead of adding the dhall package to all the dependency sets,
`cabal2nix` supports passing a flag to cabal so `cabal2nix` not only
correctly adds `dhall` to the set of dependencies but also adds a
`configureFlags` definition. This simplifies the `release.nix`.

* Remove unnecessary addition of the Dhall dependency

Missed cleanup from b21df24.