Use Data.Coerce.coerce and unboxed proxy (#91)
This should speed up protobuf serde a bit (not sure by how much), by getting rid of the container traversals to wrap things in newtypes in `encodeMessage/decodeMessage`. 

Using unboxed proxy couples us to GHC, but that's ok because we only officially support GHC 8.0+. Because this changes the API for the `Named, Finite, Primitive` classes, the version has been bumped to `0.3.0.0`.