Allow configuring Cabal flags for Nix shell
If there's a way to do this without repeating the fields and their
defaults in multiple places, let me know. Unfortunately, setting
`shell.nix` to this:

```nix
{ fast ? false
, ...
}@args:

with (import ./default.nix args);

...
```

...and entering the Nix shell like this:

```bash
$ nix-shell --argstr compiler "ghc8104"
```

...doesn't actually set the `compiler` argument to `ghc8104`. It's as
though `default.nix` considers the value from `shell.nix` to be omitted,
so it uses it's default of `ghc884`. It's really unintuitive.