Delete bogus tests
These tests all depended on `descriptor.proto` which was not valid
`proto3` syntax.  Our local copy of `descriptor.proto` was just a
vendored copy that changed `syntax = "proto2";` to `syntax = "proto3";`,
even though the syntax was not valid for `proto3`.

The only reason that this error was not caught before is because the
parser would silently ignore invalid parses (and the rest of the file).