Added a command line flag to generate large-records
Switched to a command line flag to compile-proto-file instead of a
cabal flag as suggested by @j6carey. This does not add any additional
dependency to compile-proto-file. The generated code must be compiled
with the proto3-suite library that has the `large-records` cabal flag
turned on.

The flag helps to selectively enable large-records on a per proto file
basis while compiling them.