large-records support (#225)
The large-records library is designed to avoid quadratic core code size for large records. This PR adds support to make use of this library in the code generated by `compile-proto-file`. The generated code has a different representation for the records corresponding to the messages and services. This should significantly improve the core code size and compilation timings.

See https://well-typed.com/blog/2022/03/large-records-without-the-quotes/ for an introduction to the large-records library.

The feature is controlled by a command line flag and a cabal flag:
- The command line flag `compile-proto-file --largeRecords` generates code that makes use of the `large-records` library.
- Code generated this way need to be compiled with a version of `proto3-suite` that has the `large-records` cabal flag enabled.

While the large-records representation improves the code size and compilation time of the records themselves, proto3-suite generate a large number of type class instances too. I have seen moderate improvement in memory allocation and compilation timings (around 30% and 10% respectively) compiling the proto3-suite generated code, but it is not as dramatic as one would expect. There could be additional improvements if you have some code that constructs/pattern matches these records a lot. As always, you should measure the performance before deciding to turn on this feature.

Also, note that the Generic representation provided by large-generics is different from GHC Generics. Some advanced transformations possible with GHC generics is not easy or even possible with large-generics. For example, it is possible to walk the `Rep a x` tree of GHC generics and transform it to a `Rep b x` if `b` has a similar representation to `a`. But this is not possible with large-generics.