refactor!: drop GHC 9.0 support and simplify CPP for GHC 9.2
With this commit,
support for GHC 9.0 is removed so that we can assume a minimum version of GHC 9.2.
This change allows us to simplify many CPP branches that were previously needed for older GHC versions.
The code is now cleaner and less cluttered,
as legacy conditionals have been eliminated or streamlined.