v0.9.4: Support optional fields (#289)
Older versions of the proto3 specification did not include
optional fields, but the current version recommends them.
In this change we add support for optional fields
(but in the process require proto3-wire >= 1.4.6).

This change also renames and restructures the cardinality
descriptions in the experimental `Proto3.Suite.Form` modules
to better reflect protobuf terminology and encoding.  Note
that in particular, encoding of `oneof` fields now expects
a `Maybe` or `Identity` type to reflect potential omission,
which is a breaking change to experimental usage.

And the parser now rejects attempts to put `optional`,
`repeated`, or `map` fields within a oneof.  (Technically
the old parser also rejected `optional` within `oneof`
because it rejected `optional` in general.)